telegram-reminder/requirements.txt
appdirs==1.4.4
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
filelock==3.0.12
google-api-core==1.25.1
google-auth==1.25.0
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
googleapis-common-protos==1.52.0
grpcio==1.35.0
idna==2.10
importlib-metadata==3.4.0
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-slugify==4.0.1
pytz==2021.1
requests==2.25.1
rsa==4.7
six==1.15.0
text-unidecode==1.3
typing-extensions==3.7.4.3
urllib3==1.26.3
virtualenv==20.4.2
yapf==0.30.0
zipp==3.4.0
